Events lat and long type changed to double

+++ InTheWoods/Models/Event.cs
@@ -17,8 +17,8 @@ namespace InTheWoods.Models
         public string EventDate { get; set; }
         public string EventName { get; set; }
         public string Address { get; set; }
-        public Int32 Lat { get; set; }
-        public Int32 Long { get; set; }
+        public double Lat { get; set; }
+        public double Long { get; set; }
 
     }
 }
